Integration Tests
This PR brings the "integration" tests into the extension. It extracts
the logic of the demo service that is also a good candidate for the
testing and uses the exact proto definition and the scripts we employ in
the examples.

The reason is that "integration" tests are usually hard to read, and
it's great to have the possibility to run the exact same case, just
standalone. Apart from that, they test output.