This is a port of the PR changes found here: (#25)
* This is a port of the PR changes found here:
https://github.com/grafana/k6/pull/3159

Note: The bad auth/bad certificate tests appear to be very inconsistent. In grafana/k6 I had to remove the timeout for them to pass in CI. Removing the timeout in grafana/xk6-grpc caused the tests to fail locally. This may need to be investigated or a bug raised to determine if k6 is not raising grpc connection errors on bad authentication until context deadline.